requirements.txt
# Data processing
pandas>=2.0.0
sqlalchemy>=2.0.0

# Add Flask and its dependencies
flask>=2.3.3
werkzeug>=2.3.7
jinja2>=3.1.2
itsdangerous>=2.1.2
click>=8.1.3

# RSS Parser dependencies
feedparser==6.0.10
mysql-connector-python==8.2.0

# Web Scraping dependencies
requests==2.31.0
beautifulsoup4==4.12.3

# Type hints support
typing-extensions==4.9.0

# OpenAI API
openai>=1.3.5

# Environment variables
python-dotenv==1.0.0

# Add explicit httpx version
httpx>=0.27.0 